(Week 2) refactor weight attribute types from float to double

+++ src/Bedroll.java
@@ -1,6 +1,6 @@
 public class Bedroll extends Item{
     // Constructor
-    public Bedroll(String _name, float _weight){
+    public Bedroll(String _name, double _weight){
         super(_name, _weight);
     }
 }
