+++ src/FlintAndSteel.java
@@ -1,7 +1,7 @@
 public class FlintAndSteel extends Item{
 
     // Constructor
-    public FlintAndSteel(String _name, float _weight){
+    public FlintAndSteel(String _name, double _weight){
         super(_name, _weight);
     }
 }
